+++ src/index.ts
@@ -29,6 +29,7 @@ export * from './components/box';
 export * from './components/breadcrumb';
 export * from './components/button';
 export * from './components/button-group';
+export * from './components/calendar';
 export * from './components/card';
 export * from './components/carousel';
 export * from './components/chat-bubble';